Remove the idea of a separate manual edition number
I don't believe this has ever been useful, and I doubt that we've
consistently incremented it in the past.  The reality is that the
manual evolves with the rest of Guile, so Guile's version number
should apply to the manual too.  It is conceivable that we might
one day revise the manual for an existing Guile release, so we keep
a REVISION number to cover that.