Revert "Snarfing .x depends on built headers"
This reverts commit edd6d6e2805894e21afe51276d1af65d5395180d.

As reported at <https://bugs.gentoo.org/show_bug.cgi?id=590528#c10>,
that commit failed to fix the bug.

That commit was based on the mistaken belief that the make rule syntax
".c.x: $(BUILT_INCLUDES)" means the same as ".c.x:" but with the added
prerequisites "$(BUILT_INCLUDES)".  However, as explained in section
10.7 (Old-Fashioned Suffix Rules) of the GNU Make manual:

      Suffix rules cannot have any prerequisites of their own.  If they
   have any, they are treated as normal files with funny names, not as
   suffix rules.  Thus, the rule:

        .c.o: foo.h
                $(CC) -c $(CFLAGS) $(CPPFLAGS) -o $@ $<

   tells how to make the file '.c.o' from the prerequisite file 'foo.h',
   and is not at all like the pattern rule:

        %.o: %.c foo.h
                $(CC) -c $(CFLAGS) $(CPPFLAGS) -o $@ $<

   which tells how to make '.o' files from '.c' files, and makes all
   '.o' files using this pattern rule also depend on 'foo.h'.